+++ src/server/devices/baseDeviceInterfaces/iActuator.ts
@@ -1,16 +1,24 @@
 import { ActuatorSettings } from '../../../models';
 import { iRoomDevice } from './iRoomDevice';
+import { iTemporaryDisableAutomatic } from './iTemporaryDisableAutomatic';
 
-export interface iActuator extends iRoomDevice {
+export interface iActuator extends iRoomDevice, iTemporaryDisableAutomatic {
   /**
    * The settings for this Actuator primarily for controlling its automatic actions
    */
   settings: ActuatorSettings;
 
+  targetAutomaticState: boolean;
+
   /**
    * The state value of the device
    */
   readonly actuatorOn: boolean;
+
+  /**
+   * Queued value for the actuator
+   */
+  queuedValue: boolean | null;
 
   /**
    * Persisting the current states of this device to the database